navbar UI done
