+++ frontend/src/components/auth/Signup.jsx
@@ -22,6 +22,7 @@ const Signup = () => {
 
   const handleChange = (e) => {
     setFormData({ ...formData, [e.target.name]: e.target.value });
+    console.log(formData);
   };
 
   const handleRoleChange = (role) => {
